// src/app/interfaces/common/tag.interface.ts
/* eslint-disable prettier/prettier */
export interface Tag {
  _id?: string;
  name?: string;
  nameEn?: string;
  queti?: string;
  quetiEn?: string;
  pointQuenty?: string;
  pointDescription?: string;
  pointQuentyEn?: string;
  pointDescriptionEn?: string;
  pointLink?: string;
  pointQuenty2?: string;
  pointDescription2?: string;
  pointQuenty2En?: string;
  pointDescription2En?: string;
  pointLink2?: string;
  pointQuenty3?: string;
  pointDescription3?: string;
  pointQuenty3En?: string;
  pointDescription3En?: string;
  pointLink3?: string;
  image?: string;
  createdAt?: Date;
  updatedAt?: Date;
}